requirements.txt
pandas
biopython
tqdm
